Crate revm_inspectors

revm Inspector implementations, such as call tracers

Feature Flags

js-tracer: Enables a JavaScript tracer implementation. This pulls in extra dependencies (such as boa, tokio and serde_json).

Modules

access_list

An inspector implementation for an EIP2930 Accesslist

stack

An inspector stack abstracting the implementation details of each inspector and allowing to hook on block/transaction execution, used in the main RETH executor.

tracing

An inspector for recording traces

Macros

call_inspectors

Helper macro to call the same method on multiple inspectors without resorting to dynamic dispatch