Crate revm_inspectors

revm Inspector implementations, such as call tracers

Feature Flags

js-tracer: Enables a JavaScript tracer implementation. This pulls in extra dependencies (such as boa, tokio and serde_json).

evm-inspectors

Common revm inspector implementations.

Originally part of reth as the reth-revm-inspectors crate.

License
Licensed under either of Apache License, Version 2.0 or MIT license at your option.
Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in these crates by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Modules

access_list

An inspector implementation for an EIP2930 Accesslist

stack

An inspector stack abstracting the implementation details of each inspector and allowing to hook on block/transaction execution, used in the main RETH executor.

tracing

An inspector for recording traces

Macros

call_inspectors

Helper macro to call the same method on multiple inspectors without resorting to dynamic dispatch