Module revm_inspectors::tracing

An inspector for recording traces

Modules

geth

Geth style trace builders for debug_ namespace Geth trace builder

js

Javascript inspector

parity

Parity style trace builders for trace_ namespace

types

Types for representing call trace items.

Structs

CallTraceArena

An arena of recorded traces.

FourByteInspector

Fourbyte tracing inspector that records all function selectors and their calldata sizes.

GethTraceBuilder

A type for creating geth style traces

OpcodeCountInspector

An inspector that counts all opcodes.

ParityTraceBuilder

A type for creating parity style traces

TracingInspector

An inspector that collects call traces.

TracingInspectorConfig

Gives guidance to the TracingInspector.

Enums

StackSnapshotType

How much of the stack to record. Nothing, just the items pushed, or the full stack